Address review comments to fix the dmsRecord class
RBS appendRecord/getRecord implementation

disk structure and db config change for MVCC

dis structure change for MVCC

create on disk structure (cs, cl, record) for MVCC

mvcc update meta data location

add mvcc opt list

In memory record and hash table definition

review commits

appendRecord commit1

call appendRecord during update save old ver

tested appendRecord/getRecord

interface to check page visiability

create on disk structure (cs, cl, record) for MVCC

mvcc update meta data location

In memory record and hash table definition

appendRecord commit1

logic to write in memory old version to RBS

fix defect in init, address capped record size issue, add in-flight migration issue

address review comments for disk changes part 1

add restriction for mvccon

clean up of dmsRecord for v0 vs v1

address review comment on dmsRecord

check transactionon for mvccon

remove dmsRBSSUmgr from oldversionCB

+++ SequoiaDB/engine/include/dmsScanner.hpp
@@ -168,7 +168,7 @@ namespace engine
          BOOLEAN              _firstRun ;
          BOOLEAN              _hasLockedRecord ;
          dpsTransCB           *_pTransCB ;
-         INT32                _recordLock ;
+         INT8                 _recordLock ;
          BOOLEAN              _needUnLock ;
          BOOLEAN              _CSCLLockHeld ;
          BOOLEAN              _selectForUpdate ;
